refactor: collect familiar code to ics-base crate and refactor crates of ics-conneciton and ics-channel
