+++ Makefile
@@ -14,8 +14,8 @@ ALL_CONTRACTS := \
     eth_light_client-client_type_lock \
     eth_light_client-verify_bin \
     eth_light_client-mock_business_type_lock \
-	ics-conn \
-	ics-chan \
+	ics-connection \
+	ics-channel \
 	ics-packet
 
 .PHONY: all-contracts all-contracts-in-docker
@@ -116,11 +116,11 @@ eth_light_client-verify_bin: ${OUTPUT_DIR}/eth_light_client-verify_bin
 .PHONY: eth_light_client-mock_business_type_lock
 eth_light_client-mock_business_type_lock: ${OUTPUT_DIR}/eth_light_client-mock_business_type_lock
 
-.PHONY: ics-conn
-ics-conn: ${OUTPUT_DIR}/ics-conn
+.PHONY: ics-connection
+ics-connection: ${OUTPUT_DIR}/ics-connection
 
-.PHONY: ics-chan
-ics-chan: ${OUTPUT_DIR}/ics-chan
+.PHONY: ics-channel
+ics-channel: ${OUTPUT_DIR}/ics-channel
 
 .PHONY: ics-packet
 ics-packet: ${OUTPUT_DIR}/ics-packet
