--- contracts/ics/axon_client/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "axon-client"
-version = "0.1.0"
-edition = "2021"
-license = "MIT"
-homepage = "https://github.com/synapseweb3/ibc-ckb-contracts"
-repository = "https://github.com/synapseweb3/ibc-ckb-contracts"
-
-[dependencies]
-ckb-ics-axon = { git = "https://github.com/synapseweb3/ckb-ics.git", rev = "61e8330" }
-axon-tools-riscv = { version = "0.1.1", features = ["proof"] }
-axon-types = { git = "https://github.com/axonweb3/axon-contract", rev = "8c2338a" }
-rlp = { version = "0.5.2", default-features = false }
-molecule = { version = "0.7", default-features = false }
-
-[features]
-default = ["debugging"]
-debugging = []
-
-[profile.release]
-overflow-checks = true
-lto = false
-codegen-units = 1
-strip = true
