+++ contracts/ics/base/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "ibc-ckb_contracts-ics-chan"
-version = "0.1.0-alpha.0"
+name = "ics-base"
+version = "0.1.0"
 edition = "2021"
 license = "MIT"
 homepage = "https://github.com/synapseweb3/ibc-ckb-contracts"
@@ -8,15 +8,16 @@ repository = "https://github.com/synapseweb3/ibc-ckb-contracts"
 
 [dependencies]
 ckb-std = "0.13.0"
-ckb-ics-axon = { git = "https://github.com/synapseweb3/ckb-ics.git", rev = "61e8330" }
+ckb-ics-axon = { git = "https://github.com/synapseweb3/ckb-ics.git", rev = "ef124f9" }
+axon-tools-riscv = { version = "0.1.1", features = ["proof"] }
+axon-types = { git = "https://github.com/axonweb3/axon-contract", rev = "8c2338a" }
 rlp = { version = "0.5.2", default-features = false }
-ckb-standalone-types = { version = "0.1.2", default-features = false }
 tiny-keccak = { version = "2.0.2", features = ["keccak"] }
-axon-client = { version = "*", path = "../axon_client" }
+molecule = { version = "0.7", default-features = false }
 
 [features]
-default = []
-debugging = ["axon-client/debugging"]
+default = ["debugging"]
+debugging = []
 
 [profile.release]
 overflow-checks = true
