+++ contracts/ics/base/src/axon_client.rs
@@ -1,5 +1,3 @@
-#![no_std]
-
 extern crate alloc;
 use alloc::vec::Vec;
 use axon_tools_riscv::types::{AxonBlock, Proof as AxonProof, Validator};
@@ -36,8 +34,8 @@ impl Client for AxonClient {
             .map_err(|_| VerifyError::InvalidReceiptProof)
     }
 
-    fn client_id(&self) -> &[u8] {
-        self.id.as_slice()
+    fn client_id(&self) -> &[u8; 32] {
+        &self.id
     }
 }
