+++ contracts/ics/base/src/error.rs
@@ -12,11 +12,16 @@ pub enum Error {
     Encoding,
     UnknownSysError,
 
+    UnexpectedMsg,
+
+    PacketEncoding,
+    ChannelEncoding,
     ConnectionEncoding,
     EnvelopeEncoding,
     MsgEncoding,
 
     WitnessIsIncorrect,
+    WitnessIsNotExisted,
     WitnessInputOrOutputIsNone,
     CellDataUnmatch,
 
@@ -25,11 +30,19 @@ pub enum Error {
     FailedToLoadClientId,
     FailedToCreateClient,
 
-    UnexpectedMsg,
     ConnectionProofInvalid,
+    ChannelProofInvalid,
+    PacketProofInvalid,
+
     ConnectionHashUnmatch,
-    ClientCreateWrongClientId,
+    ChannelHashUnmatch,
+    PacketHashUnmatch,
+
     ConnectionLock,
+    ChannelLock,
+    PacketLock,
+
+    ClientCreateWrongClientId,
     ClientCreateWrongConnectionCell,
 }
