+++ contracts/ics/base/src/lib.rs
@@ -1,0 +1,8 @@
+#![no_std]
+
+pub mod axon_client;
+pub mod error;
+pub mod handler;
+pub mod utils;
+
+pub use ckb_ics_axon as ckb_ics;
